Moved distribution functions to distributionController
